feat: report jsonnet module version
