+++ cmd/drydock/main.go
@@ -4,17 +4,8 @@ import (
 	"errors"
 	"fmt"
 	"os"
-	"runtime/debug"
 
 	"github.com/sholdee/drydock/internal/cli"
-)
-
-const (
-	argoCDModulePath       = "github.com/argoproj/argo-cd/v3"
-	gitOpsEngineModulePath = "github.com/argoproj/argo-cd/gitops-engine"
-	helmModulePath         = "helm.sh/helm/v4"
-	kustomizeModulePath    = "sigs.k8s.io/kustomize/api"
-	kubernetesModulePath   = "k8s.io/apimachinery"
 )
 
 var (
@@ -30,6 +21,7 @@ func main() {
 		GitOpsEngineModule: moduleLabel(gitOpsEngineModulePath),
 		HelmModule:         moduleLabel(helmModulePath),
 		KustomizeModule:    moduleLabel(kustomizeModulePath),
+		JsonnetModule:      moduleLabel(jsonnetModulePath),
 		KubernetesModule:   moduleLabel(kubernetesModulePath),
 	})
 	if err := cmd.Execute(); err != nil {
@@ -40,28 +32,4 @@ func main() {
 		fmt.Fprintln(os.Stderr, err)
 		os.Exit(2)
 	}
-}
-
-func moduleLabel(modulePath string) string {
-	info, ok := debug.ReadBuildInfo()
-	if !ok {
-		return modulePath
-	}
-	for _, dep := range info.Deps {
-		if dep.Path == modulePath {
-			return formatModuleLabel(*dep)
-		}
-	}
-	return modulePath
-}
-
-func formatModuleLabel(module debug.Module) string {
-	label := module.Path
-	if module.Version != "" {
-		label += "@" + module.Version
-	}
-	if module.Replace != nil {
-		label += " => " + formatModuleLabel(*module.Replace)
-	}
-	return label
 }
